+++ extension/lib/getShopUrl.js
@@ -1,0 +1,12 @@
+/**
+ * @param {PipelineContext} context
+ * @returns {Promise<GetShopUrlResponse>}
+ */
+module.exports = async (context) => {
+  const storeDomain = context.config.storeDomain ? context.config.storeDomain.trim() : ''
+  const domain = storeDomain || `store-${context.config.storeHash}.mybigcommerce.com`
+
+  return {
+    'url': `https://${domain}`
+  }
+}