+++ extension/lib/shopgate/customer/errors/UnexpectedResponseError.js
@@ -1,0 +1,8 @@
+class UnexpectedResponseError extends Error {
+  constructor () {
+    super()
+    this.message = 'Unexpected 200 response from Bigcommerce login action.'
+  }
+}
+
+module.exports = UnexpectedResponseError