+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- login to BigCommerce and logout of BigCommerce
+- frontend logic for opening registration url
 
 ## [0.3.0] - 2018-05-08
 ### Added